requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-cors-headers==3.2.1
django-dotenv==1.4.2
django-filter==2.2.0
django-froala-editor==3.1.0
django-jalali-date==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
itypes==1.2.0
jdatetime==3.6.2
Jinja2==2.11.2
kavenegar==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
openapi-codec==1.3.2
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
simplejson==3.17.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
